+++ com.unity.render-pipelines.lightweight/Runtime/LightweightRenderPipeline.cs
@@ -236,8 +236,9 @@ namespace UnityEngine.Rendering.LWRP
 #endif
             cameraData.postProcessEnabled = cameraData.postProcessLayer != null && cameraData.postProcessLayer.isActiveAndEnabled;
 
-            // Disables postprocessing in mobile VR. It's stable on mobile yet.
-            if (cameraData.isStereoEnabled && Application.isMobilePlatform)
+            // On Android, Postprocessing v2 works with single-pass double-wide mode and is disabled for multiview
+            var xrDesc = UnityEngine.XR.XRSettings.eyeTextureDesc;
+            if (cameraData.isStereoEnabled && Application.isMobilePlatform && Application.platform == RuntimePlatform.Android && xrDesc.dimension == TextureDimension.Tex2DArray)
                 cameraData.postProcessEnabled = false;
 
             Rect cameraRect = camera.rect;
@@ -252,8 +253,10 @@ namespace UnityEngine.Rendering.LWRP
             cameraData.renderScale = (camera.cameraType == CameraType.Game) ? cameraData.renderScale : 1.0f;
 
             bool anyShadowsEnabled = settings.supportsMainLightShadows || settings.supportsAdditionalLightShadows;
-            cameraData.maxShadowDistance = (anyShadowsEnabled) ? settings.shadowDistance : 0.0f;
-            
+            cameraData.maxShadowDistance = Mathf.Min(settings.shadowDistance, camera.farClipPlane);
+            cameraData.maxShadowDistance = (anyShadowsEnabled && cameraData.maxShadowDistance >= camera.nearClipPlane) ?
+                cameraData.maxShadowDistance : 0.0f;
+
             if (additionalCameraData != null)
             {
                 cameraData.maxShadowDistance = (additionalCameraData.renderShadows) ? cameraData.maxShadowDistance : 0.0f;